+++ src/google/adk/tools/environment/_utils.py
@@ -1,0 +1,26 @@
+# Copyright 2026 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared utilities for environment tools."""
+
+from __future__ import annotations
+
+from ._constants import MAX_OUTPUT_CHARS
+
+
+def truncate(text: str, limit: int = MAX_OUTPUT_CHARS) -> str:
+  """Truncate text to *limit* characters with a notice."""
+  if len(text) <= limit:
+    return text
+  return text[:limit] + f'\n... (truncated, {len(text)} total chars)'
